# public/js/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is lists of APIs use in mobile app.",
        "version": "1.0.0",
        "title": "Mini Aspire APIs by Krunal Palaniya"
    },
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "login user",
                "description": "",
                "operationId": "loginUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Body",
                        "description": "User login api params",
                        "schema": {
                            "$ref": "#/definitions/LogIn"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "LogIn": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                }
            }
        },
...
    }
}